NRF52 SAADC VDD INTERRUPT Example. Functions files were completed and they are ready to be tested

+++ NRF52/Examples/examples/Peripherals/SAADC_VDD_INTERRUPT/functions.h
@@ -5,8 +5,8 @@
  * @return      N/A
  *
  * @author      Manuel Caballero
- * @date        26/July/2018
- * @version     26/July/2018   The ORIGIN
+ * @date        27/July/2018
+ * @version     27/July/2018   The ORIGIN
  * @pre         N/A
  * @warning     N/A
  */
@@ -29,7 +29,7 @@ void conf_CLK    ( void );
 void conf_GPIO   ( void );
 void conf_TIMER0 ( void );
 void conf_UART   ( void );
-void conf_SAADC  ( volatile uint32_t* myADC_Result );
+void conf_SAADC  ( volatile int16_t* myADC_Result );
 
 
 /**@brief Constants.
